+++ ReeferTruck.java
@@ -19,7 +19,7 @@ public class ReeferTruck extends BasicTruck {
     }
 
     @Override
-    public int getCarryCapacity() {
+    public double getCarryCapacity() {
         return super.getCarryCapacity();
     }
