+++ Ship.java
@@ -23,7 +23,7 @@ public class Ship extends Vehicle {
 }
 
     @Override
-    public int getCarryCapacity() {
+    public double getCarryCapacity() {
         return 0;
     }
